refactor(kwin-effect): extract ctor/dtor to plasmazoneseffect/lifecycle.cpp
Step 2 of the file split per docs/kwin-effect-refactor-plan.md. Moves
PlasmaZonesEffect::PlasmaZonesEffect() (558 lines) and the destructor
(89 lines) out of plasmazoneseffect.cpp into a new translation unit at
kwin-effect/plasmazoneseffect/lifecycle.cpp.

The lc logging category remains defined once via Q_LOGGING_CATEGORY in
plasmazoneseffect.cpp; lifecycle.cpp uses Q_DECLARE_LOGGING_CATEGORY to
log under the same category without duplicating storage.

No API or behavior change. plasmazoneseffect.cpp drops from 6075 to
5429 lines.
