+++ kwin-effect/plasmazoneseffect/lifecycle.cpp
@@ -1,0 +1,688 @@
+// SPDX-FileCopyrightText: 2026 fuddlesworth
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#include "../plasmazoneseffect.h"
+
+#include <PhosphorProtocol/ClientHelpers.h>
+#include <PhosphorProtocol/ServiceConstants.h>
+#include <PhosphorProtocol/WireTypes.h>
+
+#include <effect/effecthandler.h>
+#include <core/output.h>
+#include <workspace.h>
+
+#include <QCoreApplication>
+#include <QDBusConnection>
+#include <QDBusMessage>
+#include <QDBusPendingCall>
+#include <QDBusPendingCallWatcher>
+#include <QDBusServiceWatcher>
+#include <QEvent>
+#include <QLoggingCategory>
+#include <QPointer>
+#include <QTimer>
+#include <QVarLengthArray>
+
+#include "../autotilehandler.h"
+#include "../compositorclock.h"
+#include "../dragtracker.h"
+#include "../kwin_compositor_bridge.h"
+#include "../navigationhandler.h"
+#include "../screenchangehandler.h"
+#include "../snapassisthandler.h"
+#include "../windowanimator.h"
+
+namespace PlasmaZones {
+
+// `lcEffect` is defined in plasmazoneseffect.cpp via Q_LOGGING_CATEGORY. Re-declare
+// here so this TU can log under the same category without re-defining storage.
+Q_DECLARE_LOGGING_CATEGORY(lcEffect)
+
+PlasmaZonesEffect::PlasmaZonesEffect()
+    : OffscreenEffect()
+    , m_autotileHandler(std::make_unique<AutotileHandler>(this))
+    , m_navigationHandler(std::make_unique<NavigationHandler>(this))
+    , m_screenChangeHandler(std::make_unique<ScreenChangeHandler>(this))
+    , m_snapAssistHandler(std::make_unique<SnapAssistHandler>(this))
+    // Phase 3: per-output motion clocks drive every AnimatedValue in the
+    // controller. One `CompositorClock` per `LogicalOutput` so mixed
+    // refresh-rate displays (60 Hz + 144 Hz being the common case)
+    // phase-lock independently instead of beating against a shared
+    // process-wide clock. Populated below via effects->screens() and
+    // maintained via the screenAdded/screenRemoved signals. The
+    // fallback unbound clock covers: (a) the bootstrap window before
+    // screens() populates, (b) windows whose `screen()` is null
+    // mid-migration, (c) test paths that don't drive KWin::effects.
+    , m_motionClockFallback(std::make_unique<CompositorClock>(nullptr))
+    , m_windowAnimator(std::make_unique<WindowAnimator>())
+    , m_dragTracker(std::make_unique<DragTracker>(this))
+    , m_compositorBridge(std::make_unique<KWinCompositorBridge>(this))
+{
+    PhosphorProtocol::registerWireTypes();
+
+    // Single-worker pool for off-loading user-texture loads. See the
+    // header docstring for `m_textureLoaderPool` for the rationale —
+    // serialised loads keep the dedupe cheap and avoid duplicate GPU
+    // uploads if multiple shader transitions reference the same file
+    // in quick succession.
+    m_textureLoaderPool.setMaxThreadCount(1);
+
+    // Populate per-output clocks from the currently-known output set.
+    // Subsequent hotplug events land in onScreenAdded / onScreenRemoved.
+    //
+    // Order: connect signals FIRST, then iterate the current screens()
+    // snapshot. A screen plugged in between those two steps would
+    // otherwise be missed — the signal wouldn't have an attached slot
+    // yet, and the loop would already have run. With the signals
+    // connected first, the worst case is a duplicate `onScreenAdded`
+    // call (once via signal, once via loop). `onScreenAdded` is
+    // idempotent (re-insertion check against m_motionClocksByOutput)
+    // so the duplicate is a no-op.
+    if (KWin::effects) {
+        connect(KWin::effects, &KWin::EffectsHandler::screenAdded, this, &PlasmaZonesEffect::onScreenAdded);
+        connect(KWin::effects, &KWin::EffectsHandler::screenRemoved, this, &PlasmaZonesEffect::onScreenRemoved);
+        for (KWin::LogicalOutput* output : KWin::effects->screens()) {
+            onScreenAdded(output);
+        }
+        // Seed the cursor cache with the live position so the first frame
+        // after a fresh shader install with iMouse declared sees the real
+        // cursor. The default-constructed QPointF(0, 0) would otherwise be
+        // misinterpreted as INSIDE any window whose frame contains the
+        // origin (i.e. all windows on the primary monitor with origin at
+        // (0, 0)) for one frame, producing a false-positive hover spike
+        // before prePaintScreen overwrites the cache on the next tick.
+        m_cachedCursorGlobal = KWin::effects->cursorPos();
+    }
+
+    // Wire the fallback clock as the animator's default. The animator's
+    // clockForHandle override resolves the per-output clock at
+    // startAnimation time; the default kicks in only when a window has
+    // no resolvable output (which is rare but real — XWayland
+    // bootstrap, mid-migration with a null screen()).
+    m_windowAnimator->setClock(m_motionClockFallback.get());
+    m_windowAnimator->setOutputClockResolver([this](KWin::LogicalOutput* output) -> PhosphorAnimation::IMotionClock* {
+        return clockForOutput(output);
+    });
+    m_windowAnimator->setOnAnimationCompleteCallback([this](KWin::EffectWindow* w) {
+        // Only tear down ANIMATOR-DRIVEN shader transitions
+        // (durationMs == 0; the leg rides m_windowAnimator's timeline).
+        // Time-based transitions (durationMs > 0; window.open / close /
+        // focus / etc.) have their own QTimer teardown scheduled by
+        // tryBeginShaderForEvent — without this guard, a zone.snapIn
+        // transition that's been superseded by a window.* event leaves
+        // the original animator running its geometry tween, and that
+        // animator's eventual completion would prematurely kill the
+        // successor (whose own QTimer hasn't fired yet).
+        auto it = m_shaderTransitions.find(w);
+        if (it == m_shaderTransitions.end() || it->second.durationMs > 0) {
+            return;
+        }
+        endShaderTransition(w);
+    });
+    connect(&m_animationShaderRegistry, &PhosphorAnimationShaders::AnimationShaderRegistry::effectsChanged, this,
+            [this]() {
+                QVarLengthArray<KWin::EffectWindow*, 8> windows;
+                for (auto& [w, _] : m_shaderTransitions)
+                    windows.push_back(w);
+                for (auto* w : windows)
+                    endShaderTransition(w);
+                Q_ASSERT(m_shaderTransitions.empty());
+                m_shaderCache.clear();
+                // Drop the texture cache too — a hot-reload that swaps a
+                // texture file behind the same metadata.json path needs
+                // a fresh upload to pick up the new contents. The cache
+                // is keyed by absolute path; without this clear a
+                // file-content change with no path change would never
+                // refresh.
+                //
+                // Bump the cache generation rather than draining the
+                // loader pool synchronously. `waitForDone()` on the GL
+                // thread would block the compositor for tens of ms when
+                // a worker is mid-rasterise of a 1024x1024 SVG (the
+                // worst case for `loadUserTextureImage`). Workers
+                // already in flight will complete their CPU rasterise,
+                // but their queued GL upload checks the generation
+                // captured at submission time against
+                // `m_textureCacheGeneration` and discards if mismatched
+                // — so no stale (pre-reload) bytes can re-populate the
+                // cleared cache. Clear immediately so the next
+                // `beginShaderTransition` hits the synchronous fallback
+                // path and uploads fresh content.
+                ++m_textureCacheGeneration;
+                m_textureLoadsInFlight.clear();
+                m_textureCache.clear();
+            });
+
+    // Frame-geometry shadow flush timer. Debounces per-window
+    // windowFrameGeometryChanged signals and pushes the latest geometry to
+    // the daemon at ~20Hz so daemon-local shortcut handlers (float toggle,
+    // etc.) have fresh geometry without a round-trip. Single-shot timer
+    // re-armed on each incoming change — the flush fires at most one D-Bus
+    // call per window per 50ms window regardless of how many pixels moved.
+    m_frameGeometryFlushTimer = new QTimer(this);
+    m_frameGeometryFlushTimer->setSingleShot(true);
+    m_frameGeometryFlushTimer->setInterval(50);
+    connect(m_frameGeometryFlushTimer, &QTimer::timeout, this, &PlasmaZonesEffect::flushPendingFrameGeometry);
+
+    // Connect DragTracker signals
+    //
+    // Performance optimization: keyboard grab and D-Bus dragMoved calls are deferred
+    // until an activation trigger is detected. This eliminates 60Hz D-Bus traffic and
+    // keyboard grab/ungrab overhead for non-zone window drags (discussion #167).
+    connect(m_dragTracker.get(), &DragTracker::dragStarted, this,
+            [this](KWin::EffectWindow* w, const QString& windowId, const QRectF& geometry) {
+                qCDebug(lcEffect) << "Window move started -" << w->windowClass()
+                                  << "current modifiers:" << static_cast<int>(m_currentModifiers);
+
+                // Note: `cursor.drag` is intentionally NOT wired here. The
+                // OffscreenEffect pipeline operates on window content; firing
+                // a shader at drag start through it is indistinguishable from
+                // `window.move`, and synchronously colliding with the
+                // `windowStartUserMovedResized` lambda's `window.move` install
+                // means whichever fires second wins (it would be `window.move`
+                // here). See `ProfilePaths::CursorDrag` doc comment — the path
+                // is reserved for a future cursor-decoration / drag-shadow
+                // surface.
+
+                // Fire beginDrag async to get a daemon-authoritative policy.
+                // While the reply is pending, we
+                // default m_currentDragPolicy to a conservative snap-path so
+                // the worst case (stale effect cache would have said autotile
+                // but daemon knows better, or vice-versa) is a brief overlay
+                // flash rather than a dead drag. The reply handler flips the
+                // bypass flag retroactively a few ms later if the daemon says
+                // this is an autotile drag.
+                //
+                // This replaces the previous stale-cache read of
+                // m_autotileHandler->isAutotileScreen() as the single source
+                // of truth for drag-start routing — root cause of the
+                // post-settings-reload dead-drag window found in #310 log
+                // forensics.
+                m_currentDragPolicy = PhosphorProtocol::DragPolicy{};
+                m_currentDragPolicy.streamDragMoved = true;
+                m_currentDragPolicy.showOverlay = true;
+                m_currentDragPolicy.grabKeyboard = true;
+                m_currentDragPolicy.captureGeometry = true;
+
+                const QString startScreenId = getWindowScreenId(w);
+                const QRect frame = geometry.toRect();
+                QDBusMessage beginMsg = QDBusMessage::createMethodCall(
+                    PhosphorProtocol::Service::Name, PhosphorProtocol::Service::ObjectPath,
+                    PhosphorProtocol::Service::Interface::WindowDrag, QStringLiteral("beginDrag"));
+                beginMsg << windowId << frame.x() << frame.y() << frame.width() << frame.height() << startScreenId
+                         << static_cast<int>(m_currentMouseButtons);
+                QDBusPendingCall beginPending = QDBusConnection::sessionBus().asyncCall(beginMsg);
+                auto* beginWatcher = new QDBusPendingCallWatcher(beginPending, this);
+                QPointer<KWin::EffectWindow> safeW = w;
+                const QString capturedWindowId = windowId;
+                const QString capturedScreenId = startScreenId;
+                connect(
+                    beginWatcher, &QDBusPendingCallWatcher::finished, this,
+                    [this, safeW, capturedWindowId, capturedScreenId](QDBusPendingCallWatcher* bw) {
+                        bw->deleteLater();
+                        QDBusPendingReply<PhosphorProtocol::DragPolicy> reply = *bw;
+                        if (!reply.isValid()) {
+                            qCWarning(lcEffect) << "beginDrag reply invalid:" << reply.error().message();
+                            return;
+                        }
+                        const PhosphorProtocol::DragPolicy policy = reply.value();
+                        if (const QString err = policy.validationError(); !err.isEmpty()) {
+                            qCWarning(lcEffect) << "beginDrag reply rejected:" << err
+                                                << "— keeping conservative snap-path policy for" << capturedWindowId;
+                            return;
+                        }
+                        m_currentDragPolicy = policy;
+                        qCInfo(lcEffect) << "beginDrag reply:" << capturedWindowId
+                                         << "bypass=" << m_currentDragPolicy.bypassReason
+                                         << "stream=" << m_currentDragPolicy.streamDragMoved
+                                         << "immediateFloat=" << m_currentDragPolicy.immediateFloatOnStart;
+                        // If the daemon confirms autotile, flip the effect
+                        // state to bypass mode. Usually the effect-side
+                        // fast path below already did this synchronously;
+                        // this catches the stale-cache case where the fast
+                        // path missed.
+                        if (m_currentDragPolicy.bypassReason == PhosphorProtocol::DragBypassReason::AutotileScreen) {
+                            if (!m_dragBypassedForAutotile) {
+                                m_dragBypassedForAutotile = true;
+                                m_dragBypassScreenId = capturedScreenId;
+                                qCInfo(lcEffect) << "beginDrag: retroactive autotile bypass for" << capturedWindowId;
+                            }
+                            // Apply immediate float transition if the policy
+                            // says so and the window wasn't already floated
+                            // by the fast path. Using QPointer so we skip
+                            // if the window was destroyed between drag-start
+                            // and reply.
+                            if (safeW && m_currentDragPolicy.immediateFloatOnStart
+                                && !isWindowFloating(capturedWindowId)
+                                && !m_dragFloatedWindowIds.contains(capturedWindowId)) {
+                                m_autotileHandler->handleDragToFloat(safeW, capturedWindowId, capturedScreenId,
+                                                                     /*immediate=*/true);
+                                m_dragFloatedWindowIds.insert(capturedWindowId);
+                            }
+                        }
+                    });
+
+                // Fast path: the effect-side autotile cache is USUALLY correct.
+                // We still consult it synchronously so the common case runs at
+                // zero latency. The async beginDrag reply above runs as a
+                // correction layer for the cases where the cache is stale
+                // (post-settings-reload — the #310 scenario).
+                if (m_autotileHandler->isAutotileScreen(startScreenId)) {
+                    m_dragBypassedForAutotile = true;
+                    m_dragBypassScreenId = startScreenId;
+                    // Reorder mode: the daemon owns drag-insert preview for tile
+                    // swapping. Skip the synchronous float transition — we want
+                    // the tile to stay visually in place while the daemon runs
+                    // moveToTiledPosition on each cursor tick. The effect still
+                    // flips into bypass state so snap-path logic is suppressed.
+                    const bool reorderMode = m_cachedAutotileDragBehavior == EffectAutotileDragBehavior::Reorder;
+                    // If the window is currently autotile-tiled, restore its
+                    // title bar and pre-autotile size NOW (synchronously, during
+                    // the interactive move). This mirrors snap mode, where
+                    // dragging a snapped window out of its zone visibly restores
+                    // the free-floating size before release — without this, the
+                    // user drags a borderless tile-sized window and only sees it
+                    // become a floating window after they drop.
+                    //
+                    // Guarded on isTrackedWindow so we don't touch windows that
+                    // are already floating (not in the autotile tree).
+                    if (!reorderMode && m_autotileHandler->isTrackedWindow(windowId) && !isWindowFloating(windowId)) {
+                        m_autotileHandler->handleDragToFloat(w, windowId, m_dragBypassScreenId, /*immediate=*/true);
+                        // Mark as drag-floated so the daemon's pre-tile geometry
+                        // restore (applyGeometryForFloat, triggered by the
+                        // setWindowFloatingForScreen call at drop) is skipped in
+                        // slotApplyGeometryRequested — the window should stay
+                        // where the user drops it, not snap back to a stored rect.
+                        m_dragFloatedWindowIds.insert(windowId);
+                    }
+                    return;
+                }
+                m_dragBypassedForAutotile = false;
+                m_dragActivationDetected = false;
+                m_dragStartedSent = false;
+                m_pendingDragWindowId = windowId;
+                m_pendingDragGeometry = geometry;
+                m_snapDragStartScreenId = getWindowScreenId(w);
+
+                // beginDrag already initialized daemon-side snap-drag state
+                // (called internally from the adaptor). The effect only needs
+                // to decide whether to grab the keyboard for local Escape
+                // handling.
+                detectActivationAndGrab();
+                // Grab keyboard to intercept Escape before KWin's MoveResizeFilter.
+                // Without this, Escape cancels the interactive move AND the overlay.
+                // With the grab, Escape only dismisses the overlay while the drag continues.
+                if (!m_keyboardGrabbed) {
+                    KWin::effects->grabKeyboard(this);
+                    m_keyboardGrabbed = true;
+                }
+            });
+    connect(
+        m_dragTracker.get(), &DragTracker::dragMoved, this, [this](const QString& windowId, const QPointF& cursorPos) {
+            // Cross-VS flip detection is daemon-owned. The
+            // daemon's updateDragCursor handler computes policy at the
+            // cursor position and emits dragPolicyChanged when it flips.
+            // The effect reacts via slotDragPolicyChanged (see below).
+            //
+            // Here we only forward the cursor to the daemon as a
+            // fire-and-forget call. The daemon-side dispatch handles
+            // both the snap-path overlay updates and the cross-VS
+            // detection in a single round trip.
+
+            // In autotile bypass — skip snap zone processing locally;
+            // the daemon's updateDragCursor still watches for a flip
+            // BACK to snap mode.
+            const bool bypassed = m_currentDragPolicy.bypassReason == PhosphorProtocol::DragBypassReason::AutotileScreen
+                || m_dragBypassedForAutotile;
+            if (!bypassed) {
+                // Gate D-Bus calls on activation trigger state so a drag
+                // without any intent to use zones doesn't flood the bus
+                // at 30Hz. This is a local input-event optimization; it
+                // isn't policy and doesn't come from the daemon.
+                if (!detectActivationAndGrab() && !m_cachedZoneSelectorEnabled && m_triggersLoaded) {
+                    return;
+                }
+            }
+
+            // Forward the cursor to the daemon. For snap drags, this
+            // drives overlay/zone detection. For bypass drags, the
+            // daemon watches the cursor for a cross-VS flip and emits
+            // dragPolicyChanged when the policy changes.
+            PhosphorProtocol::ClientHelpers::fireAndForget(
+                this, PhosphorProtocol::Service::Interface::WindowDrag, QStringLiteral("updateDragCursor"),
+                {windowId, qRound(cursorPos.x()), qRound(cursorPos.y()), static_cast<int>(m_currentModifiers),
+                 static_cast<int>(m_currentMouseButtons)},
+                QStringLiteral("updateDragCursor"));
+        });
+    connect(m_dragTracker.get(), &DragTracker::dragStopped, this,
+            [this](KWin::EffectWindow* w, const QString& windowId, bool cancelled) {
+                // Release keyboard grab before handling drag end
+                if (m_keyboardGrabbed) {
+                    KWin::effects->ungrabKeyboard();
+                    m_keyboardGrabbed = false;
+                }
+
+                // Clear the drag-floated marker on every drag end. Historically
+                // this marker was used to suppress a post-drag pre-tile geometry
+                // restore (applyGeometryForFloat), but the current daemon-side
+                // drag-end path goes through AutotileEngine::setWindowFloat →
+                // windowFloatingStateSynced → syncAutotileFloatStatePassive,
+                // which never emits applyGeometryForFloat. Leaving the marker
+                // set after a drag leaks it into subsequent Meta+F toggles:
+                // the next user float is silently skipped, the window's visual
+                // position diverges from the daemon's shadow, and then a
+                // float→tile toggle overwrites the stored pre-tile rect with
+                // the stale tile zone — permanently corrupting the restore
+                // target (#bug: zed/firefox/plasmazones-settings resize issues).
+                m_dragFloatedWindowIds.remove(windowId);
+
+                // Single entry point for drag-end dispatch. The
+                // daemon owns the decision; callEndDrag sends endDrag and
+                // the reply handler applies whatever PhosphorProtocol::DragOutcome comes back
+                // (ApplySnap / ApplyFloat / RestoreSize / NoOp / etc.).
+                //
+                // The autotile branch special-casing that used to live here
+                // is gone — cross-VS transitions were applied mid-drag by
+                // slotDragPolicyChanged, and final drop-time actions are
+                // encoded in the PhosphorProtocol::DragOutcome.
+                callEndDrag(w, windowId, cancelled);
+
+                // Clear drag state for the next session.
+                m_currentDragPolicy = PhosphorProtocol::DragPolicy{};
+                m_dragBypassedForAutotile = false;
+                m_dragBypassScreenId.clear();
+                m_snapDragStartScreenId.clear();
+                m_dragActivationDetected = false;
+                m_dragStartedSent = false;
+                m_pendingDragWindowId.clear();
+                m_pendingDragGeometry = QRectF();
+            });
+
+    // Connect to window lifecycle signals
+    connect(KWin::effects, &KWin::EffectsHandler::windowAdded, this, &PlasmaZonesEffect::slotWindowAdded);
+    connect(KWin::effects, &KWin::EffectsHandler::windowClosed, this, &PlasmaZonesEffect::slotWindowClosed);
+
+    // Belt-and-suspenders: windowClosed removes animations, but if a deferred
+    // timer re-adds one between windowClosed and windowDeleted, the Item tree
+    // will be torn down while an animation entry still references the window.
+    // Purge here to prevent SIGSEGV in animationBounds → expandedGeometry.
+    // Also clean up caches that slotWindowClosed may have already cleared —
+    // QHash::take/remove on missing keys is a no-op, so this is safe.
+    connect(KWin::effects, &KWin::EffectsHandler::windowDeleted, this, [this](KWin::EffectWindow* w) {
+        endShaderTransition(w);
+        m_windowAnimator->removeAnimation(w);
+        if (m_windowIdCache.contains(w)) {
+            const QString cachedId = m_windowIdCache.take(w);
+            m_windowIdReverse.remove(cachedId);
+        }
+        m_trackedScreenPerWindow.remove(w);
+        // Drop per-window shader-event bookkeeping. m_lastFocusShaderWindow is
+        // a QPointer that auto-nulls on destroy, so it's already cleaned up;
+        // m_lastFullyMaximized is a raw-pointer-keyed QHash so we explicitly
+        // erase here to keep it bounded across long sessions.
+        m_lastFullyMaximized.remove(w);
+    });
+
+    connect(KWin::effects, &KWin::EffectsHandler::windowActivated, this, &PlasmaZonesEffect::slotWindowActivated);
+
+    // Update the daemon's primary screen override when KDE Display Settings change
+    if (auto* ws = KWin::Workspace::self()) {
+        connect(ws, &KWin::Workspace::outputOrderChanged, this, [this]() {
+            auto* workspace = KWin::Workspace::self();
+            if (workspace && m_daemonServiceRegistered) {
+                const auto outputs = workspace->outputOrder();
+                if (!outputs.isEmpty()) {
+                    PhosphorProtocol::ClientHelpers::fireAndForget(
+                        this, PhosphorProtocol::Service::Interface::Screen, QStringLiteral("setPrimaryScreenFromKWin"),
+                        {outputs.first()->name()}, QStringLiteral("setPrimaryScreenFromKWin"));
+                }
+            }
+        });
+    }
+
+    // mouseChanged is the only reliable way to get modifier state in a KWin effect on Wayland;
+    // QGuiApplication::queryKeyboardModifiers() doesn't work since effects run in the compositor.
+    connect(KWin::effects, &KWin::EffectsHandler::mouseChanged, this, &PlasmaZonesEffect::slotMouseChanged);
+
+    // Connect to screen geometry changes for keepWindowsInZonesOnResolutionChange feature
+    // In KWin 6, use virtualScreenGeometryChanged (not per-screen signal)
+    connect(KWin::effects, &KWin::EffectsHandler::virtualScreenGeometryChanged, m_screenChangeHandler.get(),
+            &ScreenChangeHandler::slotScreenGeometryChanged);
+    // Invalidate screen ID cache and refresh virtual screen definitions on screen changes
+    // (connector names may be reassigned, physical screen geometry changes invalidate
+    // virtual screen absolute geometry)
+    connect(KWin::effects, &KWin::EffectsHandler::virtualScreenGeometryChanged, this, [this]() {
+        m_screenIdCache.clear();
+        m_lastEffectiveScreenId.clear();
+    });
+
+    // Connect to daemon's settingsChanged D-Bus signal
+    QDBusConnection::sessionBus().connect(PhosphorProtocol::Service::Name, PhosphorProtocol::Service::ObjectPath,
+                                          PhosphorProtocol::Service::Interface::Settings,
+                                          QStringLiteral("settingsChanged"), this, SLOT(slotSettingsChanged()));
+    qCInfo(lcEffect) << "Connected to daemon settingsChanged D-Bus signal";
+
+    // Connect to virtual screen changes — daemon emits this when a physical screen's
+    // virtual subdivisions are added, removed, or modified.
+    QDBusConnection::sessionBus().connect(PhosphorProtocol::Service::Name, PhosphorProtocol::Service::ObjectPath,
+                                          PhosphorProtocol::Service::Interface::Screen,
+                                          QStringLiteral("virtualScreensChanged"), this,
+                                          SLOT(onVirtualScreensChanged(QString)));
+
+    // Connect to keyboard navigation D-Bus signals
+    connectNavigationSignals();
+
+    // Connect to autotile D-Bus signals
+    m_autotileHandler->connectSignals();
+    m_autotileHandler->loadSettings();
+
+    // Verify daemon availability asynchronously to avoid blocking the compositor.
+    // CRITICAL: Do NOT use synchronous isServiceRegistered() here. The daemon
+    // registers its D-Bus service name in init() BEFORE start() runs heavy
+    // initialization and BEFORE the event loop begins (main.cpp:88→94→102).
+    // During that window, isServiceRegistered() returns true but the daemon
+    // can't process messages. Any synchronous QDBusInterface creation would
+    // trigger Introspect, blocking KWin for up to the D-Bus timeout (~25s).
+    //
+    // Instead, send an async Introspect — if the daemon responds, it's fully
+    // operational and we trigger slotDaemonReady(). If it can't respond (still
+    // initializing), the call times out harmlessly and we wait for the
+    // daemonReady D-Bus signal instead.
+    {
+        QDBusMessage introspect = QDBusMessage::createMethodCall(
+            PhosphorProtocol::Service::Name, PhosphorProtocol::Service::ObjectPath,
+            QStringLiteral("org.freedesktop.DBus.Introspectable"), QStringLiteral("Introspect"));
+        auto* watcher = new QDBusPendingCallWatcher(QDBusConnection::sessionBus().asyncCall(introspect, 3000), this);
+        connect(watcher, &QDBusPendingCallWatcher::finished, this, [this](QDBusPendingCallWatcher* w) {
+            w->deleteLater();
+            QDBusPendingReply<QString> reply = *w;
+            if (reply.isValid() && !m_daemonServiceRegistered) {
+                // Daemon responded — it's fully operational.
+                // Trigger the same ready flow as the daemonReady signal.
+                slotDaemonReady();
+            }
+        });
+    }
+
+    // Connect to daemon's daemonReady signal — emitted at the end of Daemon::start()
+    // after all initialization is complete and the daemon can process D-Bus messages.
+    // This is the safe point to set m_daemonServiceRegistered and create QDBusInterfaces.
+    QDBusConnection::sessionBus().connect(PhosphorProtocol::Service::Name, PhosphorProtocol::Service::ObjectPath,
+                                          PhosphorProtocol::Service::Interface::LayoutRegistry,
+                                          QStringLiteral("daemonReady"), this, SLOT(slotDaemonReady()));
+
+    // Watch for daemon D-Bus service registration and unregistration.
+    // After a daemon restart, m_lastCursorOutput is still valid in the effect
+    // but the daemon's lastCursorScreenName/lastActiveScreenName are empty.
+    // Without this, keyboard shortcuts (rotate, etc.) operate on all screens
+    // because resolveShortcutScreen returns nullptr.
+    //
+    // On Wayland, this watcher uses D-Bus monitoring (not X11 selection),
+    // which works reliably across both sessions.
+    auto* serviceWatcher = new QDBusServiceWatcher(
+        PhosphorProtocol::Service::Name, QDBusConnection::sessionBus(),
+        QDBusServiceWatcher::WatchForRegistration | QDBusServiceWatcher::WatchForUnregistration, this);
+    connect(serviceWatcher, &QDBusServiceWatcher::serviceUnregistered, this, [this]() {
+        qCInfo(lcEffect) << "Daemon service unregistered";
+        m_daemonServiceRegistered = false;
+        // Also clear the bridge-registration in-flight gate. Without
+        // this, a daemon-restart racing the in-flight registerBridge
+        // reply leaves the gate set: the new daemon's `daemonReady`
+        // signal arrives, slotDaemonReady sees the gate true and
+        // bails, and the gate only clears later when the stale call's
+        // error reply arrives — by which time no further signal will
+        // re-trigger slotDaemonReady. The effect would sit idle
+        // indefinitely. Resetting here keeps the gate authoritative
+        // across daemon restarts.
+        m_bridgeRegistrationInFlight = false;
+        m_daemonReadyRestoresDone = false;
+        m_daemonReadyWindowStateProcessed = false;
+        m_snapRestoreCache.clear();
+
+        // Restore borderless and monocle-maximized windows — daemon state is gone
+        m_autotileHandler->restoreAllBorderless();
+        m_autotileHandler->restoreAllMonocleMaximized();
+        clearAllBorders();
+    });
+    connect(serviceWatcher, &QDBusServiceWatcher::serviceRegistered, this, [this]() {
+        qCInfo(lcEffect) << "Daemon registered: waiting for daemonReady signal";
+
+        // DO NOT set m_daemonServiceRegistered = true here.
+        // The daemon registers its D-Bus service name in init(), BEFORE start()
+        // runs heavy initialization and BEFORE the event loop begins. Keep the
+        // flag false until the daemon's own daemonReady signal fires (end of
+        // Daemon::start()), confirming it can handle D-Bus requests.
+
+        // Reconnect daemonReady signal — Qt may cache the old daemon's unique bus
+        // name in match rules, so refresh for the new daemon instance.
+        // Disconnect first to prevent duplicate match rules (Qt doesn't deduplicate),
+        // which would cause slotDaemonReady to fire twice on the same signal.
+        QDBusConnection::sessionBus().disconnect(PhosphorProtocol::Service::Name, PhosphorProtocol::Service::ObjectPath,
+                                                 PhosphorProtocol::Service::Interface::LayoutRegistry,
+                                                 QStringLiteral("daemonReady"), this, SLOT(slotDaemonReady()));
+        QDBusConnection::sessionBus().connect(PhosphorProtocol::Service::Name, PhosphorProtocol::Service::ObjectPath,
+                                              PhosphorProtocol::Service::Interface::LayoutRegistry,
+                                              QStringLiteral("daemonReady"), this, SLOT(slotDaemonReady()));
+    });
+
+    // NOTE: syncFloatingWindowsFromDaemon() and loadCachedSettings() are NOT
+    // called here. m_daemonServiceRegistered is false at this point (set only by
+    // slotDaemonReady), so any ensureInterface() call would bail out immediately.
+    // All daemon state sync is deferred to slotDaemonReady().
+
+    // Connect to existing windows
+    const auto windows = KWin::effects->stackingOrder();
+    for (KWin::EffectWindow* w : windows) {
+        setupWindowConnections(w);
+    }
+
+    // The daemon disables KWin's Quick Tile via kwriteconfig6. We don't reserve electric borders
+    // here because that would turn on the edge effect visually; the daemon's config approach
+    // is the right way to prevent Quick Tile from activating.
+
+    // Seed m_lastCursorOutput with the compositor's active screen. This ensures
+    // the daemon has a valid cursor screen even if no mouse movement occurs after login.
+    // slotMouseChanged will overwrite this as soon as the cursor moves.
+    //
+    // The actual D-Bus push to the daemon happens in slotDaemonReady(), which fires
+    // either from the async Introspect callback above (daemon already running) or
+    // from the daemonReady D-Bus signal (daemon starts later). We do NOT push here
+    // to avoid synchronous QDBusInterface creation on the compositor thread.
+    auto* initialScreen = KWin::effects->activeScreen();
+    if (initialScreen) {
+        m_lastCursorOutput = initialScreen->name();
+    }
+
+    qCInfo(lcEffect) << "initialized: C++ effect with D-Bus support and mouseChanged connection";
+}
+
+PlasmaZonesEffect::~PlasmaZonesEffect()
+{
+    // Sever the registry's `effectsChanged` connection BEFORE anything
+    // else runs. The slot lambda touches `m_shaderTransitions`,
+    // `m_shaderCache`, and `m_textureCache` — all declared AFTER
+    // `m_animationShaderRegistry` in the header (h:507 vs h:698+), so
+    // they destruct FIRST in C++ reverse-declaration order. The
+    // registry destructs LAST, and any signal it (or its underlying
+    // file-watcher) emits during its own member teardown would
+    // dispatch to the slot AFTER the cache members are gone — UAF.
+    // Disconnect now while everything is still alive.
+    disconnect(&m_animationShaderRegistry, nullptr, this, nullptr);
+
+    // Drain the texture loader pool before any other teardown. A
+    // worker that's mid-rasterise would otherwise post a queued
+    // upload via QMetaObject::invokeMethod against `this` AFTER our
+    // members start tearing down — UAF on the texture cache.
+    // QThreadPool::waitForDone is called here BEFORE we let the
+    // member's destructor run (which itself blocks on waitForDone)
+    // so the pending invokeMethod posts are flushed against a still-
+    // intact `this`.
+    //
+    // Distinct from the hot-reload path in the `effectsChanged`
+    // lambda above: hot-reload bumps `m_textureCacheGeneration` (no
+    // wait) so workers in flight can discard their queued upload
+    // without blocking the compositor. Shutdown REQUIRES the wait —
+    // the queued uploads need to be flushed against a live `this`,
+    // not raced against member destruction.
+    m_textureLoaderPool.waitForDone();
+    m_textureLoadsInFlight.clear();
+
+    // Restore borderless and monocle-maximized windows so they recover properly.
+    // Guard against compositor teardown — effects may outlive the stacking order.
+    if (KWin::effects) {
+        m_autotileHandler->restoreAllBorderless();
+        m_autotileHandler->restoreAllMonocleMaximized();
+        clearAllBorders();
+    }
+
+    if (m_keyboardGrabbed && KWin::effects) {
+        // Symmetric with the `if (KWin::effects)` guard above: during
+        // compositor teardown KWin::effects can be null, and an
+        // unguarded deref here would crash even though we reached the
+        // destructor body cleanly. The compositor's own teardown
+        // releases the grab when KWin::effects is gone.
+        KWin::effects->ungrabKeyboard();
+        m_keyboardGrabbed = false;
+    }
+    m_screenChangeHandler->stop();
+    // We no longer reserve/unreserve edges; the daemon disables KWin snap via config.
+
+    // Explicitly tear down every active shader transition before the
+    // member-by-member destruction runs. `endShaderTransition` calls
+    // `setShader(window, nullptr)` and `unredirect(window)` to release
+    // KWin's offscreen state cleanly; relying on default destruction
+    // would let the offscreen FBOs linger until KWin's effect-removed
+    // bookkeeping ran. Iterates a snapshot because endShaderTransition
+    // erases from `m_shaderTransitions` mid-loop.
+    //
+    // Guarded by `if (KWin::effects)` matching the clearAllBorders /
+    // ungrabKeyboard guards above: during compositor teardown the global
+    // is null and `endShaderTransition` dereferences it (setShader,
+    // unredirect, refWindow). The compositor's own teardown reclaims
+    // the offscreen state when KWin::effects is gone.
+    if (KWin::effects) {
+        QVarLengthArray<KWin::EffectWindow*, 8> activeWindows;
+        for (auto& [w, _] : m_shaderTransitions) {
+            activeWindows.push_back(w);
+        }
+        for (auto* w : activeWindows) {
+            endShaderTransition(w);
+        }
+        // endShaderTransition queues each window's `unrefWindow` via
+        // QMetaObject::invokeMethod(this, ..., Qt::QueuedConnection) to
+        // avoid use-after-free in paintWindow's expiry fall-through. In
+        // the destructor path that defer is unsafe in the opposite
+        // direction: ~QObject runs after this body returns and discards
+        // pending posted MetaCalls targeted at `this`, so the queued
+        // unrefs would never fire and KWin's EffectWindow refcount
+        // stays incremented for every close-grab transition active at
+        // teardown — leaking the close grab. Drain the queue here, while
+        // `this` is still fully constructed and the lambdas can safely
+        // run. The lambdas only call `KWin::effects->unrefWindow(...)`
+        // and don't touch our member state after the unref, so
+        // synchronous dispatch from the dtor body is sound.
+        QCoreApplication::sendPostedEvents(this, QEvent::MetaCall);
+    }
+}
+
+} // namespace PlasmaZones
